function/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-identity
azure-storage-blob
openai
requests
aiohttp>=3.9.0  # For async HTTP requests
Pillow>=10.0.0  # For robust image validation
